Update chronos.py - model.device (#11)
*Issue #, if available:* N/A

*Description of changes:*

Thanks for the very clean impl of the Model, Tokenizer, and Pipeline. 

I was curios about it and found a minor improvement in the API - what do
you think about it? Feel free to close. Change is untested.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.